+++ spade_llm/demo/platform/contractnet/discovery.py
@@ -15,7 +15,7 @@ from spade_llm.core.behaviors import MessageHandlingBehavior, MessageTemplate
 from spade_llm.core.conf import configuration, Configurable
 
 _ID_SEPARATOR = "/"
-DF_AGENT = "DF"
+DF_ADRESS = "df"
 logger = logging.getLogger(__name__)
